Move and rename and split one final  time
