+++ test/corecel/CMakeLists.txt
@@ -30,6 +30,7 @@ celeritas_add_test(OpaqueId.test.cc)
 
 # Cont
 celeritas_add_test(cont/Array.test.cc)
+celeritas_add_test(cont/LdgSpan.test.cc)
 celeritas_add_test(cont/InitializedValue.test.cc)
 celeritas_add_test(cont/Span.test.cc)
 celeritas_add_test(cont/LabelIdMultiMap.test.cc)
